refact(Team): split team classes by file
