+++ lib/nba_rb/team/player_on_off_summary.rb
@@ -1,0 +1,15 @@
+module NbaRb
+  module Team
+    class PlayerOnOffSummary < Dashboard
+      @endpoint = 'teamplayeronoffsummary'
+
+      def on_court
+        create_stats_hash(@data[1])
+      end
+
+      def off_court
+        create_stats_hash(@data[2])
+      end
+    end
+  end
+end
