+++ lib/nba_rb/team/year_over_year_splits.rb
@@ -1,0 +1,11 @@
+module NbaRb
+  module Team
+    class YearOverYearSplits < Dashboard
+      @endpoint = 'teamdashboardbyyearoveryear'
+
+      def by_year
+        create_stats_hash(@data[1])
+      end
+    end
+  end
+end
