+++ lib/nba_rb/version.rb
@@ -1,3 +1,3 @@
 module NbaRb
-  VERSION = "0.1.1"
+  VERSION = "0.2.0"
 end
